# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DrawResult {
  key: string;
  tier: number;
  tierName: string;
  remaining: number;
}

export interface KeyStats {
  tier1: number;
  tier2: number;
  tier3: number;
  total: number;
}

export interface ErrorResponse {
  error: string;
}