# lib/api-zod/src/generated/types/keyStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface KeyStats {
  tier1: number;
  tier2: number;
  tier3: number;
  total: number;
}